# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Toy Mall Inventory Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: string;
  name: string;
  sku: string;
  /** @nullable */
  barcode?: string | null;
  category: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /** @nullable */
  purchasePrice?: number | null;
  /**
   * ISO datetime string after which the sale price is no longer valid
   * @nullable
   */
  salePriceUntil?: string | null;
  stock: number;
  lowStockThreshold: number;
  createdAt: string;
}

export interface CreateProductInput {
  name: string;
  sku: string;
  /** @nullable */
  barcode?: string | null;
  category: string;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  /** @nullable */
  purchasePrice?: number | null;
  /**
   * ISO datetime string after which the sale price is no longer valid
   * @nullable
   */
  salePriceUntil?: string | null;
  stock?: number;
  lowStockThreshold?: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  supplierId?: string | null;
}

export interface UpdateProductInput {
  name?: string;
  sku?: string;
  /** @nullable */
  barcode?: string | null;
  category?: string;
  price?: number;
  /** @nullable */
  salePrice?: number | null;
  /** @nullable */
  purchasePrice?: number | null;
  /**
   * ISO datetime string after which the sale price is no longer valid
   * @nullable
   */
  salePriceUntil?: string | null;
  stock?: number;
  lowStockThreshold?: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  supplierId?: string | null;
}

export type StockUpdateBodyType =
  (typeof StockUpdateBodyType)[keyof typeof StockUpdateBodyType];

export const StockUpdateBodyType = {
  IN: "IN",
  OUT: "OUT",
  ADJUSTMENT: "ADJUSTMENT",
} as const;

export interface StockUpdateBody {
  type: StockUpdateBodyType;
  quantity: number;
  userId?: string;
}

export type StockLogType = (typeof StockLogType)[keyof typeof StockLogType];

export const StockLogType = {
  IN: "IN",
  OUT: "OUT",
  ADJUSTMENT: "ADJUSTMENT",
  RETURN: "RETURN",
} as const;

export interface StockLog {
  id: string;
  productId: string;
  productName: string;
  productSku: string;
  type: StockLogType;
  quantity: number;
  /** @nullable */
  userId?: string | null;
  createdAt: string;
}

export interface Sale {
  id: string;
  productId: string;
  productName: string;
  productSku: string;
  quantity: number;
  totalPrice: number;
  createdAt: string;
}

export interface StockUpdateResponse {
  product: Product;
  log: StockLog;
  sale?: Sale;
}

export interface QrCodeResponse {
  sku: string;
  url: string;
  qrDataUrl: string;
}

export interface DashboardSummary {
  totalProducts: number;
  totalStock: number;
  totalStockValue: number;
  lowStockCount: number;
  todayInCount: number;
  todayOutCount: number;
}

export interface TodayActivity {
  inCount: number;
  outCount: number;
  inQuantity: number;
  outQuantity: number;
}

export interface CategoryStat {
  category: string;
  productCount: number;
  totalStock: number;
  stockValue: number;
}

export interface CheckoutItem {
  productId: string;
  quantity: number;
  price: number;
  mrp?: number;
}

/**
 * Mode of payment. `credit` records the bill as unpaid (receivable); requires `customerPhone` so the debtor is identifiable.
 */
export type CheckoutInputPaymentMode =
  (typeof CheckoutInputPaymentMode)[keyof typeof CheckoutInputPaymentMode];

export const CheckoutInputPaymentMode = {
  cash: "cash",
  upi: "upi",
  credit: "credit",
} as const;

/**
 * How to interpret the discount value
 */
export type CheckoutInputDiscountType =
  (typeof CheckoutInputDiscountType)[keyof typeof CheckoutInputDiscountType];

export const CheckoutInputDiscountType = {
  percent: "percent",
  amount: "amount",
} as const;

export interface CheckoutInput {
  items: CheckoutItem[];
  /** Mode of payment. `credit` records the bill as unpaid (receivable); requires `customerPhone` so the debtor is identifiable. */
  paymentMode: CheckoutInputPaymentMode;
  customerName?: string;
  customerPhone?: string;
  /** Raw discount value (e.g. 10 for 10% or 200 for ₹200) */
  discount?: number;
  /** How to interpret the discount value */
  discountType?: CheckoutInputDiscountType;
}

export type BillPaymentStatus =
  (typeof BillPaymentStatus)[keyof typeof BillPaymentStatus];

export const BillPaymentStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface Bill {
  id: string;
  billNumber?: number;
  totalAmount: number;
  /** Cumulative amount collected against this bill. */
  amountPaid: number;
  paymentStatus: BillPaymentStatus;
  itemsCount: number;
  paymentMode: string;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  discount?: number | null;
  /** @nullable */
  discountType?: string | null;
  createdAt: string;
}

export interface BillItem {
  id: string;
  productId: string;
  productName: string;
  productSku: string;
  quantity: number;
  price: number;
  /** @nullable */
  mrp?: number | null;
  subtotal: number;
}

export interface BillDetail {
  bill: Bill;
  items: BillItem[];
}

export interface CheckoutResponse {
  bill: Bill;
  items: BillItem[];
}

/**
 * How the customer paid this installment
 */
export type RecordPaymentInputPaymentMode =
  (typeof RecordPaymentInputPaymentMode)[keyof typeof RecordPaymentInputPaymentMode];

export const RecordPaymentInputPaymentMode = {
  cash: "cash",
  upi: "upi",
} as const;

export interface RecordPaymentInput {
  /** Amount being collected now (clamped to outstanding balance) */
  amount: number;
  /** How the customer paid this installment */
  paymentMode?: RecordPaymentInputPaymentMode;
}

export interface Debtor {
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  outstanding: number;
  billCount: number;
  lastBillAt: string;
}

export interface ReceivablesSummary {
  totalOutstanding: number;
  billCount: number;
  topDebtors: Debtor[];
}

export type ListProductsParams = {
  /**
   * Search by name or SKU
   */
  search?: string;
  /**
   * Filter by category
   */
  category?: string;
  /**
   * Filter to low stock products only
   */
  lowStock?: boolean;
};

export type ListStockLogsParams = {
  /**
   * Filter by product ID
   */
  productId?: string;
  /**
   * Filter by log type
   */
  type?: ListStockLogsType;
  /**
   * Only return logs created today (server time)
   */
  today?: boolean;
  /**
   * Number of logs to return
   */
  limit?: number;
  /**
   * Offset for pagination
   */
  offset?: number;
};

export type ListStockLogsType =
  (typeof ListStockLogsType)[keyof typeof ListStockLogsType];

export const ListStockLogsType = {
  IN: "IN",
  OUT: "OUT",
  ADJUSTMENT: "ADJUSTMENT",
  RETURN: "RETURN",
} as const;

export type ListSalesParams = {
  /**
   * Filter by product ID
   */
  productId?: string;
  limit?: number;
  offset?: number;
};