# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Toy Mall Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or SKU
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: lowStock
          in: query
          schema:
            type: boolean
          description: Filter to low stock products only
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        category:
          type: string
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        purchasePrice:
          type: ["number", "null"]
        salePriceUntil:
          type: ["string", "null"]
          description: ISO datetime string after which the sale price is no longer valid
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - sku
        - category
        - price
        - stock
        - lowStockThreshold
        - createdAt
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        category:
          type: string
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        purchasePrice:
          type: ["number", "null"]
        salePriceUntil:
          type: ["string", "null"]
          description: ISO datetime string after which the sale price is no longer valid
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        imageUrl:
          type: ["string", "null"]
        supplierId:
          type: ["string", "null"]
      required:
        - name
        - sku
        - category
        - price
...